Added Firebase functionality
imported firebase libraries, `gson`. Created a new class called `FireBaseHandler`, which handles interaction with the firebase that has functions to register and login to a user. Used this register function to get a POC of actually registering in the register menu. included some changes in the UI because of this change.