Added login and user data retrieval, fixed UI
Added an option to login to the application and a function for getting the user data from a user object, which will be used when logging in or registering. When the user is logged in and in the login screen, it will bring him to the home screen with his information automatically (for that reason, I also made the logout function actually log out of the user). also fixed the UI as the login is with an email, not a username.