Refactored objects & populated arrays pre-fill
Changed the item and data objects to fit the post object from the RT database.
Added a new function `getAllPosts` to get all the posts from the RT database and a function `populateArrays` to populate the arrays in the `myData` class.
The home fragment now uses this function to populate the arrays before filling in the data.