Enhanced UI and added location/distance features
Added emojis to represent fields in TextViews.
Added an option to get a full name from a user ID with a function `getFullName` and used this function to set the full name in the post preview.
Added a function to calculate the distance between the current location and a target location `setDistanceToLocation` and used this function to update the location in the post preview.
Changed the default post preview texts to a text that represents when they are loading.