Fixed address bar, deprecated classes and bugs
When you add parking, your current address will be shown in the address bar, and when you pick a new address, the address bar will be updated accordingly.
Fix a lot of deprecated classes and changed the code accordingly: `ViewPager`->`ViewPager2`, `FragmentPagerAdapter` -> `FragmentStateAdapter`, use of `fusedLocationClient`
A lot of bug fixes and code quality improvements.