Fixed & optimized post updates, RecyclerView bug
Added static lists that keeps track of all `CustomAdapter`s and `Post`s; those lists will be used to make sure that when one post is updated, all of them will be updated.
Instead of getting the user posts twice, we are now only getting the posts list once and using the `postID` list to filter them; this also fixed a bug with retrieving user `postID`s.
Fixed a bug when liking a post in parking; the information won't update in the `RecyclerView`s.